Adding a game session
