+++ main.py
@@ -1,6 +1,9 @@
-def main() -> None:
-    print("Hello from time-think!")
+from session import run
+
+
+def main() -> str:
+    return run()
 
 
 if __name__ == "__main__":
-    main()
+    print(main())
