--- test_main.py
@@ -1,2 +1,0 @@
-def test_sum() -> None:
-    assert True
